feat - add UploadUpdateStore & separate store using slice pattern
