+++ store/upload/useUploadStore.ts
@@ -1,0 +1,16 @@
+import { UploadStoreState } from '#types/storeType/upload';
+import create from 'zustand';
+import { persist } from 'zustand/middleware';
+
+import { createUploadSlice } from './UploadSlice';
+
+export const useUploadStore = create(
+  persist<UploadStoreState>(
+    (...a) => ({
+      ...createUploadSlice(...a),
+    }),
+    {
+      name: 'upload-storage',
+    },
+  ),
+);
