+++ store/upload/useUploadUpdateStore.ts
@@ -1,0 +1,11 @@
+import create from 'zustand';
+
+import { createUpdateSlice, UpdateSlice } from './UpdateSlice';
+import { createUploadSlice, UploadSlice } from './UploadSlice';
+
+export const useUploadUpdateStore = create<UpdateSlice & UploadSlice>(
+  (...a) => ({
+    ...createUploadSlice(...a),
+    ...createUpdateSlice(...a),
+  }),
+);
